+++ import-service/package.json
@@ -10,7 +10,11 @@
   "license": "ISC",
   "description": "hometask for practising AWS cloud services",
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.600.0",
+    "@aws-sdk/client-sqs": "^3.600.0",
+    "@smithy/protocol-http": "4.0.1",
     "aws-sdk": "^2.1642.0",
-    "csv-parser": "^3.0.0"
+    "csv-parser": "^3.0.0",
+    "uuid": "^10.0.0"
   }
 }
