--- import-service/products.csv
@@ -1,9 +1,0 @@
-id;title;description;price;count
-1;Product_1;ProductOne;10;2
-2;Product_2;ProductTwo;20;3
-3;Product_3;ProductThree;30;4
-4;Product_4;ProductFour;40;5
-5;Product_5;ProductFive;50;6
-6;Product_6;ProductSix;60;7
-7;Product_7;ProductSeven;70;8
-8;Product_8;ProductEight;80;9
