+++ import-service/products1.csv
@@ -1,0 +1,11 @@
+id;title;description;price;count
+a5e12a86-9399-4741-9639-a728168b8d54;Product_01;ProductOne;10;2
+b570c7e4-6ed4-4f25-89d8-653f335afdbb;Product_02;ProductTwo;20;3
+f9d9e37b-69c9-4ef5-a1c2-fc5f2fc8a235;Product_03;ProductThree;30;4
+9b3f0b1f-ebf4-47a2-a79c-7a900d4a8fa2;Product_04;ProductFour;40;5
+d016b3b1-f0c2-43cd-ba6a-208bb491a29c;Product_05;ProductFive;50;6
+53f08c3d-8d87-4e92-991d-15db4b8cbf38;Product_06;ProductSix;60;7
+d9e2b6a0-f5e3-4f07-bd45-0e9c84ae1f06;Product_07;ProductSeven;70;8
+81f3f666-e027-4a69-b865-b2c2047b8e08;Product_08;ProductEight;80;9
+7ae9a69b-35ee-48c1-a90f-d1f678765346;Product_09;ProductNine;90;10
+ce788d24-e34d-46bf-abb8-ae77f8dfd486;Product_10;ProductTen;100;11
