+++ import-service/products2.csv
@@ -1,0 +1,11 @@
+id;title;description;price;count
+ff71e63b-381e-40c0-bcef-d3e1f49c5f2b;Product_11;ProductEleven;110;12
+1f3b475f-f5b8-43de-8a22-2b55a0e2b22a;Product_12;ProductTwelve;120;13
+9bf0e61d-1a46-4a7c-8dae-91f6d8b725b3;Product_13;ProductThirteen;130;14
+84f2928a-7b0c-45b8-a92d-4062f6ea621a;Product_14;ProductFourteen;140;15
+cef8a14e-657b-4844-b02e-2ff3e5c8b20e;Product_15;ProductFifteen;150;16
+4ab2c940-e5b2-4c15-aed0-59e0db713b30;Product_16;ProductSixteen;160;17
+db24dbad-fad3-4ebc-b2f5-6c40ecfbd5ee;Product_17;ProductSeventeen;170;18
+0725d294-16fd-4878-86a1-f5ef5769cbb2;Product_18;ProductEighteen;180;19
+2091f59e-c1ad-4ad2-b9ec-7c2d21b23399;Product_19;ProductNineteen;190;20
+a49b77a5-488e-4c67-baac-e537955bd7e5;Product_20;ProductTwenty;200;21
