+++ import-service/products3.csv
@@ -1,0 +1,11 @@
+id;title;description;price;count
+4a6c3f7a-4a9e-4e07-ae8a-e4802a92ab32;Product_21;ProductTwentyOne;210;22
+1ad113c4-9bf2-45ef-a8e2-f5f4a6d4a24e;Product_22;ProductTwentyTwo;220;23
+5a84a2c1-4885-4a5e-8422-4b3b5cb82377;Product_23;ProductTwentyThree;230;24
+7b1820d1-f76c-47cb-9831-ef5d68d287c3;Product_24;ProductTwentyFour;240;25
+c5b7197c-25b1-478f-aa0f-485f82158bfd;Product_25;ProductTwentyFive;250;26
+5a9870d8-f52a-4343-afa7-4196d4d07f64;Product_26;ProductTwentySix;260;27
+ab1e1c8e-5702-4f09-a9ee-f0a6a3c4f262;Product_27;ProductTwentySeven;270;28
+b4db1569-1cf3-4acc-a17f-9f6b54deb570;Product_28;ProductTwentyEight;280;29
+f5d4ef5e-65a2-451f-9937-a04228459b10;Product_29;ProductTwentyNine;290;30
+1ea5d9f7-749b-4a80-8c1d-8bbf1745e213;Product_30;ProductThirty;300;31
