+++ links.md
@@ -1,0 +1,5 @@
+## API Gateways
+
+GET products:
+
+<https://edeupbwwm0.execute-api.eu-central-1.amazonaws.com/dev/products>
